+++ modules/app-ui/src/commonMain/kotlin/org/archivekeep/app/desktop/ui/designsystem/layout/views/constants.kt
@@ -4,10 +4,6 @@ import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.ui.unit.dp
 
-val ViewPadding = 24.dp
-val ViewItemSpacing = 24.dp
-val ViewExtraPaddingForScrollbar = 8.dp
-
 val ScrollbarPadding = PaddingValues(end = 8.dp, top = 8.dp, bottom = 8.dp)
 
 val SectionCardShape = RoundedCornerShape(11.dp, 11.dp, 9.dp, 9.dp)