+++ modules/app-ui/src/commonMain/kotlin/org/archivekeep/app/desktop/ui/views/storages/components/StoragesList.kt
@@ -26,6 +26,7 @@ import org.archivekeep.app.desktop.ui.designsystem.sections.SectionCardBottomLis
 import org.archivekeep.app.desktop.ui.designsystem.sections.SectionCardTitle
 import org.archivekeep.app.desktop.ui.designsystem.sections.sectionCardHorizontalPadding
 import org.archivekeep.app.desktop.ui.designsystem.sections.sectionCardItem
+import org.archivekeep.app.desktop.ui.designsystem.theme.AppTheme
 import org.archivekeep.app.desktop.ui.views.storages.StoragesViewState
 import org.archivekeep.utils.loading.Loadable
 
@@ -37,8 +38,8 @@ fun StoragesList(
     LoadableGuard(allStoragesLoadable) { allStorages ->
         VerticalGrid(
             columns = SimpleGridCells.Adaptive(minSize = 240.dp),
-            horizontalArrangement = Arrangement.spacedBy(14.dp),
-            verticalArrangement = Arrangement.spacedBy(16.dp),
+            horizontalArrangement = Arrangement.spacedBy(AppTheme.dimens.gridSpacingHorizontal),
+            verticalArrangement = Arrangement.spacedBy(AppTheme.dimens.gridSpacingVertical),
         ) {
             if (allStorages.isEmpty()) {
                 EmptySectionCard(emptyText)