+++ modules/app-ui/src/desktopMain/kotlin/org/archivekeep/app/desktop/ui/designsystem/theme/AppDimensions.desktop.kt
@@ -1,0 +1,6 @@
+package org.archivekeep.app.desktop.ui.designsystem.theme
+
+import androidx.compose.ui.unit.Dp
+import androidx.compose.ui.unit.dp
+
+internal actual fun viewExtraPaddingForScrollbar(): Dp = 8.dp